# openapi.yaml
openapi: 3.0.3
info:
  title: User Service API
  version: "1.0.0"
paths:
  /users/register:
    post:
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
  /auth/login:
    post:
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
  /users/me:
    get:
      summary: Получить свой профиль
      responses:
        '200':
          description: Данные профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
    put:
      summary: Обновить свой профиль
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Успешное обновление
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
components:
  schemas:
    UserCreate:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        email:
          type: string
    UserUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
        phone_number:
          type: string
        email:
          type: string
    UserOut:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
        phone_number:
          type: string
    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string